various typos and bug fixes
git-svn-id: svn://192.168.1.10/public/CoffeeMud@19083 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Commands/ListCmd.java
@@ -4546,10 +4546,10 @@ public class ListCmd extends StdCommand
 						x=CMParms.indexOf(Ability.FLAG_DESCS,str.toUpperCase().substring(5).trim());
 					if(x >= 0)
 					{
-						final int mBit = (int)CMath.pow(2, x);
+						final long mBit = CMath.pow(2, x);
 						title = (Ability.FLAG_DESCS[x]+" "+title).trim();
 						enumA = new FilteredEnumeration<Ability>(enumA, new Filterer<Ability>() {
-							final int mask = mBit;
+							final long mask = mBit;
 							@Override
 							public boolean passesFilter(final Ability obj)
 							{
